Factor auth endpoint into endpoints/auth directory.
