+++ api/globals/globals.py
@@ -2,7 +2,6 @@ from flask_api import FlaskAPI
 from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
 from config.config import Config
-from auth.auth import auth_blueprint
 import os
 
 
@@ -13,8 +12,9 @@ db = SQLAlchemy(app)
 migrate = Migrate(app, db)
 
 
+from endpoints.auth.auth import auth_endpoint_blueprint
 from endpoints.snails.snails import snails_endpoint_blueprint
 
 
-app.register_blueprint(auth_blueprint)
+app.register_blueprint(auth_endpoint_blueprint)
 app.register_blueprint(snails_endpoint_blueprint)
